Refactor mcp tool compile
