+++ src/vellum/workflows/nodes/displayable/bases/inline_prompt_node/node.py
@@ -45,12 +45,14 @@ from vellum.workflows.nodes.displayable.bases.base_prompt_node import BasePrompt
 from vellum.workflows.nodes.displayable.bases.utils import process_additional_prompt_outputs
 from vellum.workflows.outputs import BaseOutput
 from vellum.workflows.types import MergeBehavior
-from vellum.workflows.types.definition import DeploymentDefinition
+from vellum.workflows.types.definition import DeploymentDefinition, MCPServer
 from vellum.workflows.types.generics import StateType, is_workflow_class
 from vellum.workflows.utils.functions import (
     compile_function_definition,
     compile_inline_workflow_function_definition,
+    compile_mcp_tool_definition,
     compile_workflow_deployment_function_definition,
+    get_mcp_tool_name,
 )
 from vellum.workflows.utils.pydantic_schema import normalize_json
 
@@ -140,6 +142,16 @@ class BaseInlinePromptNode(BasePromptNode[StateType], Generic[StateType]):
                     normalized_functions.append(compile_inline_workflow_function_definition(function))
                 elif callable(function):
                     normalized_functions.append(compile_function_definition(function))
+                elif isinstance(function, MCPServer):
+                    tool_definitions = compile_mcp_tool_definition(function)
+                    for tool_def in tool_definitions:
+                        normalized_functions.append(
+                            FunctionDefinition(
+                                name=get_mcp_tool_name(tool_def),
+                                description=tool_def.description,
+                                parameters=tool_def.parameters,
+                            )
+                        )
                 else:
                     raise NodeException(
                         message=f"`{function}` is not a valid function definition",
